fix[hypr]: Fix neovim shortcut closing window
When we close a neovim instance with `:q` we want to go back to the terminal.

Previously doing a `:q` on a neovim instance opened with the shortcut would close the whole window. Now it goes back to a terminal.

We first open a terminal then we open a neovim instance inside it so that we can go back to the terminal once neovim is closed.